Added School and WebScraper models
and made School a ForeignKey of many school-based models
