--- django_project/core/models/__init__.py
@@ -1,2 +1,0 @@
-# WIP: from .schools import WebScraper, School
-from .courses import Professor, Term, Subject, Course, Section, Favorite
